test_api_service/requirements.txt
Django==4.2.13
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0